[2/3] Lazy-Tcl: infrastructure
This patch adds the infrastructure to make on-demand loading of Tcl scripts
possible, but it doesn't yet use this feature. That's something for the next
patch.

The core of the on-demand loading mechanism is the built-in Tcl 'unknown' proc.
Whenever the Tcl interpreter is asked to execute a command with a name that
doesn't exist, it instead executes the command with the name 'unknown' (if that
exists) and passes the original command (plus arguments) as arguments to this
'unknown' command.

So this patch implements an 'unknown' proc. This proc will lookup the name
of the script that defines the yet-unknown proc, execute that script and then
re-execute the original command. We can't magically know which script defines
which procs (without being as slow as parsing all scripts) so scripts that
want to be loaded on-demand first have to register themselves via the new
'register_lazy' proc. (And of course this registration cannot be done in the
script itself).

Next to procs Tcl scripts can also define help-texts and tab-completion
routines. So querying help or performing tab-completing on registered lazy
routines also causes the associated script to be loaded. A special case is the
'about' command: this command will search through the help text of all
commands. So this will cause all not-yet-loaded scripts to be loaded.

In the openMSX console we highlight and tab-complete proc names. We still want
this to work for names of not-yet-loaded procs. So next to the (pure Tcl)
extension described in the above two paragraphs, there are also 2 minor
extensions in the console code on the c++ side.

The previous patch already mentioned that some scripts (e.g. keybindings.tcl)
cannot be loaded on demand. To determine which script should or should not be
loaded on startup, I've introduced the convention that script with a filename
that has a leading '_' character are not automatically loaded at startup.
Hopefully this convention is good enough for backwards compatibility. E.g. if a
user has added his own scripts to his share/scripts directory, those scripts
should continue to work without any changes. Or in other words, scripts have
to opt-in for lazy loading.