Split reverse.tcl in two, one part can be loaded on-demand
The reverse.tcl script is (was) a relatively big script. So on systems where
reverse is not enabled by default (like Dingoo) it's nice if we can lazily load
this script. Though one of the things the script does is actually checking
whether reverse should be enabled or not, and this check cannot be done lazily.

The solution is to split the script in two:
- One (small) part that creates the 'auto_enable_reverse' setting, checks it
  value and optionally enables reverse (and the reversebar).
- One (big) part that contains various reverse-related helper procs and the
  implementation of the reversebar.

The (old) reverse.tcl script was the 2nd most expensive non-lazy Tcl script.
After this patch it loads 5x faster (of course only when the 2nd part isn't
actually loaded as well). By far the most expensive script is load_icons.tcl.
Not because it's a particularly big script, but because it has to load several
.png files.