Ported Y8950/YMF262 fix from MAME
The Y8950 (msx-audio), YMF262 (moonsound) and YM2413 (msx-music) all have 2
bits to select the key-scale-level (ksl). MAME recently committed a fix that
changes the interpretation of these bits:
  http://mame.dorando.at/svn/?rev=24448
These bits select between 0.0, 1.5, 3.0 or 6.0 dB/oct. Though instead of
mapping the order [0..3] to {0, 1.5, 3, 6}, the fix in MAME changed the order
to {0, 3, 1.5, 6}. The sound emulation in MAME and openMSX is originally based
on the same code. So it's likely the same fix needs to be applied to openMSX as
well.

Though because this is such an illogical order I decided to verify this on real
hardware. For Y8950 and YMF262 I could indeed verify that the fix is correct
(so emulation was wrong). However for YM2413 the original order was correct (so
MAME fix is wrong).

Looking back at the chips datasheets:
- YM2413: I only have very bad scans of this datasheet, the details of the ksl
  table are unreadable.
- Y8950: (I have a paper version) When strictly interpreting the ksl table
  (including all labels), the info is wrong. Though it's also constructed in a
  very confusing way.
- YMF262: This documentation is fine. It very clearly shows the correct (but
  weird) order.
So it's not surprising that the original sound emulation cores got this detail
wrong.

I've also reported this to the MAME developers.