[1/15] Remove 'rhythm' from Burczynski, make more similar to Okazaki
This patch removes the 'rhythm' variable from the Burczynski core. This
variable is 'always' in sync with bit 5 of register 0x0E. Though it was used to
detect when a write to reg 0x0E flipped this bit. Instead we now keep the old
value of the register as a local variable (not anymore as a member variable).

This patch also makes the code that handles writes to register 0x0E more
similar between the two cores.

The Okazaki core already didn't have a 'rhythm' member variable. Instead it
detected rhythm-mode changes by keeping track of a 'patch_number' variable per
channel. A value >= 16 indicates the channel is in rhythm-mode. A later patch
in this series will remove this 'patch_number' variable. So this patch already
changes the Okazaki core to not depend on it in the rhythm-mode-changing code.