[2/15] Removed 'instvol_r' from YM2413Burczynski
Removed the per channel 'instvol_r' member variable from the Burczynski core.
This variable is 'always' in sync with the register value '0x30+channel'.

Similar to the previous patch, this member was used to detect changes on
register writes. The Burczynski core used this as an optimization to not
re-initialize the instrument data when the upper 4 bits of register 0x30-0x38
remain the same on a write (the Okazaki core doesn't have such an
optimization).

I kept this optimization (I didn't measure if it's a useful optimization). But,
just like in the previous patch, instead of using a member variable I now use a
local variable to detect changes.