[3/15] Removed 'patch_number' from YM2413Okazaki
The Okazaki core kept a 'patch_number' member variable per channel. Though the
same information is also present in the upper 4 bits of the registers
0x30-0x38. Though in case of rhythm-mode the upper 3 channels use a 'special'
patch, that was also reflected in the 'patch_number' variable.

The first patch in this series already removed the rhythm-mode-detection usage
of this variable. Another use of it was to detect whether a channel has
selected the user-defined instrument. But this test can easily be rewritten by
testing register values.

The last use of 'patch_number' was in the de-serialization code to re-initialize
the correct instrument data. But again with slightly more control logic the same
information can be obtained by looking at the (restored) register values.