[6/15] Don't store patchFlags in YM2413Okazaki
Each channel in the Okazaki core stored a 'patchFlags' variable. This is a
combination of several boolean values from the 'Patch' data of the mod and car
slots. So it is redundant data. This patch doesn't store it in a variable
anymore, but instead calculates it each time.

After this patch the only remaining state in YM2413Okazaki::Channel are the
two 'car' and 'mod' slots.

'patchFlags' is used during sound generation to select a more specialized (and
often _much_ faster) version of the inner loop. The calculation of the
'patchFlags' variable is not part of the inner loop, but still it's used
100x-1000x per second. It's a simple but not a trivial calculation. Though
later patches in this series will mitigate this cost somewhat.

Later there will also be patches in this series that improve performance, I
promise ;-). Overall this series is performance neutral.