[9/15] Pack AM and PM fields together
The 'AM' and 'PM' boolean were extracted from the same byte. Later their only
use is to be combined again as part of the 'patchFlags' value. (In the past
these two were used separately).

So this patch keeps 'AM' and 'PM' together. The order of the bits in
'patchFlags' has changed because of this, but that's OK.