[10/15] Transform Patch::ML field
Instead of storing a value ML=[0..15] store the transformed value
'mlTable[ML]'. This is useful because the only use of ML was in this
transformation and the transformed value is likely used more often that the
number of times ML is changed.

This required updating the table generator because the 'mlTable' that used to
be local in the updatePG() method is now global.

This patch also renames:
    setFeedbackShift() -> setFB()
    setKeyScaleRate()  -> setKR()
This is more in line with the rest of this patch series.