[13/15] Transform TL + tllTable
The YM2413Okazaki::Patch::TL variable is only used as an index in a
two-dimensional array. TL is likely much more frequently used than it is
modified. So it's faster to instead of the TL value store a pointer that can be
used as a one-dimensional array.

Because TL was originally the 2nd dimension of tllTable, this array also had to
be transposed. And at the same time, because tllTable only contains small values,
I changed the type from 'unsigned' to 'byte'. Both these transformations should
be good for cache-locality.