[14/15] Transform WF, remove sintbl
YM2413Okazaki::Slot has members 'patch' and 'sintbl'. At any time is
'waveform[patch.WF] == sintbl'. So one of 'patch.WF' or 'sintbl' is redundant.
I choose to remove 'sintbl' and transform 'WF' into 'waveform[WF]' (this is
more in line with the other patches in this series).

This also allowed to remove the updateWF() method. Its only purpose was to
update 'sintbl' after 'patch.WF' was changed.