[4/15] Removed freq member variable from YM2413Okazaki
All channels in the Okazaki core keep track of their current frequency. This
value is always in sync with (12 bits from) the register values. And it's
relatively cheap to calculate this value from the registers. So I removed this
redundant variable and instead added a method that calculates it.

This frequency value is needed to update several internal other values on
register changes. It's not needed on actual sound generation. So this new
getFreq() method is called 100-1000x per second but not 1000000x, this should
be OK. (Yes, some of the routines in the inner most loops in the YM2413
emulation are called almost a million times per second, so there even these
details like calculate vs store in a variable do matter.)

To be 100% correct I should have increased the serialization-class-version of
YM2413Okazaki::Channel. Though because I already increased it in this patch
series (and I intended to commit this series in one go) I didn't increase it
again.