[5/15] Don't serialize patches[0] in YM2413Okazaki
'patches[0]' can be re-calculated from registers 0x00-0x07 on load, so no need
to (redundantly) serialize it. The other indices in the 'patches' array are
read-only, so they already weren't serialized.

The Patch::serialize() method is no longer used, so it's removed now. This will
be very useful for later patches in this series (no need to add backwards-
compatibility code to this method).