Transplanted LFO-PM from Burczynski to Okazaki
Copied to LFO-PM algorithm from the Burczynski core to the Okazaki core. I
don't know if this is exactly how the real hardware works. Though the
Burczynski code has comments that (parts of) it is tested on a real YM2413.
It's definitely so that the Burczynski algorithm is much more suited to be
implemented as a (cheap) hardware circuit than the (old) Okazaki one.

The new LFO_PM period is 8 x 1024 samples. At 3579545Hz / 72 = 49.7kHz that
corresponds to 6.1Hz. But the YM2413 datasheet says it should be 6.4Hz. I'm
currently inclined to believe the datasheet is wrong (or it assumes a different
input clock). There are more mistakes in the datasheet. The modulation depth
is roughly the same between the old and the new algorithm. The YM2413 datasheet
doesn't mention a value for the modulation depth.

I did tweak the algorithm a bit, but it should give exactly the same result. I
shifted some calculations from the sound generation routine to the
register-update routine. I couldn't measure any significant performance
difference between the (old) Okazaki and the (new) Burczynski algorithm.