Allow 0kB sample ram in MSX-AUDIO, fix Toshiba_HX-MU900
When not specifying a <sampleram> tag in a MSX-AUDIO config file, it defaults
to 256kB sample ram. Arguably this is a bad default. I did some digging through
git history and AFAIC it has always been like this. We may wish to change this
default (to 0kB) in a follow-up patch, needs discussion.

Anyway, this illogical default value caused the Toshiba_HX-MU900 to have 256kB
(instead of 0kB) sample ram. Also AFAIC this has always been like this. I added
an explicit '<sampleram>0</sampleram>' tag to fix this. IMHO this is better
than relying on some implicit default value.

There's was a (harmless) assert triggered with sampleram=0, this patch also
fixes that.