Fixed 'rom_info' Tcl script
Internally 'rom_info' used 'guess_title::guess_rom_title'. This proc is defined
in '_rom_info.tcl', but this script is lazily loaded. The script 'lazy.tcl'
contains a list of all procs names that can be lazily loaded, together with the
filename of the script that defines that proc. When I initially wrote that
script I only included all exported procs (I still think this is a good idea).
Though 'rom_info' directly called a proc inside the guess_title namespace. The
namespace is not important, but the fact that I only added exported functions
in 'lazy.tcl' is.

To fix this I:
 - added guess_rom_title in lazy.tcl (required)
 - made guess_rom_title an exported proc (optional)