Fixed joystick state after loading a savestate.
We found that after loading a savestate on Android, the joystick inputs
wouldn't respond anymore. Quibus tracked this regression down to revision
bb839473.

This revision improves the joystick emulation to listen to pin8. If this pin is
in a certain state then the joystick shouldn't return input values. After a
loadstate, the value of this pin wasn't restored correctly. So although this
patch did trigger the bug, it did not really cause it. The actual cause was
introduced in revision ddd5a27f. This changed the order in which the PSG and
joystick ports were (de)serialized. We already had code to re-write the value
of the joystick output pins on loadstate. But after rev ddd5a27f this value was
written before the correct joystick devices were re-plugged.

So once the problem was analyzed it was easy to fix: re-write the output values
after the devices are re-plugged. Because writing a value requires access to
the current EmuTime, (the largest) part of this patch adds functionality to
query this time.