Fixed Tcl argument escaping bug in ::unknown
Execute the following command in the console:
	unmute_channels Panasoft\ WS-M004\ FMPAC
The first time this gave an error:
	Unknown sounddevice: Panasoft
Simply re-executing the same command works without problems. Note that when you
use tab-completion to type the first command, the problem doesn't show up. I
discovered this problem while working on wxcatapult. Wxcatapult does send such
commands to openMSX.

The problem was located in the lazy Tcl loading mechanism. More specifically in
the part that re-executes the original command after the script has been
loaded. It went wrong when parameters had embedded space characters. It was
caused by (to me) unexpected escaping rules for the 'uplevel' Tcl command. I
verified all usages of 'uplevel' in our scripts and they all seem fine (except
for the one that I fixed in this patch).