Fixed Tcl command-verifier (e.g. on set_bp) and lazy loading
A command like
	debug set_condition {[vpeek 0] == 0}
was no longer accepted. This was because we do some basic verification on the
given expression (and command) (useful to catch simple typos). And this failed
when the script that defines the 'vpeek' proc was not yet loaded.

Note that this bug did not trigger when interactively typing this command, only
when source'ing a script that contained this command. (Though this was actually
caused by another bug. This is also fixed in this patch).

Fixed by reusing the 'isProc' from TclParser (already used by the console to
correctly highlight names of not yet loaded procs). Reusing common code is a
good idea anyway.

There have been a lot of (small) lazy-Tcl loading related bugs lately. It seems
that implementing this feature fully correct was (is?) more difficult than I
originally thought. Though I still think it's worth the effort because it
resulted in measurable faster startup performance on Dingoo (~20%) with also
less memory usage (~1MB).