Slight CPU emulation speedup: put PC at offset=0
The PC register is the most often accessed member variable in the CPUCore
class. For each emulated instruction, this variable is accessed at least twice
(read + write), many instructions access the PC register even more. If we place
the PC variable as the very first variable in the CPUCore class, then it can be
accessed by only dereferencing the 'this-pointer' (so without first having to
add an offset to the pointer).

Practically all modern CPU architectures have addressing modes that allow to
load/store data from a register+offset in a single instruction. So at first
sight the above transformation may not seem like a win. Though on x86 if the
offset is zero, then the load/store instruction can be encoded shorter. Also
keep in mind that the Z80 PC is a 16-bit value and 16-bit load/store
instructions on modern CPUs are typically more limited than their 32-bit
counter-parts. So also on non- x86 CPUs this transformation may have benefits
(but I didn't test).

There is one hidden problem: simply putting the PC variable as the first
variable in CPUCore does not put it at offset zero. At least not when the class
has virtual methods (because then vptr is put at offset=0). So this patch also
makes CPUCore non-virtual and instead moves the dynamic dispatching to the
MSXCPU class. Because we only need to dispatch between two possible classes
(Z80 and R800) this is a reasonable transformation.

Benchmark results, only modest improvements:
- code size got about 4kB smaller
- cpu emulation got about 1% faster (on x86-64)