Remove CPU base class
The previous patch made CPUCore non-virtual. Because of this, the base class
(called CPU) became almost empty. The only remaining stuff was the definition
of the registers (inside the CPURegs class). Though the code structure becomes
simpler if we remove the CPU base class and instead inherit from CPURegs.

An alternative could be to not inherit from CPURegs, but instead have it as a
member variable. I choose inheritance because it allowed many small syntactic
simplifications in the implementation of the CPUCore members (and this code
is already quite complex, so every simplification helps).

Another alternative is to also remove CPURegs and instead integrate it into
CPUCore directly. This would even allow to better optimize the difference in
the handling of the 'R' register between Z80 and R800. I didn't pick this
alternative because the debugger code currently still needs to access the
registers in a z80/r800-independent way, and this is only a very small
optimization anyway (only the 'LD A,R' and 'LD R,A' instructions on R800 would
benefit).