Speedup VDP command execution / re-enable cmdtiming=broken
Revision 618ccb implemented a cycle accurate VDP command engine. But it also
caused a massive slowdown in the emulation speed of the commands. Revision
a8451e improved the emulation speed a lot, but it still remained quite a bit
slower than the old code. Rev 618ccb also disabled the 'cmdtiming=broken'
functionality.

This patch re-enables 'cmdtiming=broken' and at the same time speeds up the
command engine emulation speed. Though the speed is not back at the original
level. (It can certainly still be improved, but I doubt it's possible to get
all speed back).

Some very quick benchmark results:
- command engine emulation speed is improved by a factor ~1.7x
- very much depending on what and how exactly you measure this translates to
  2%-10% faster total emulation speed (for code that heavily uses VDP commands)

The main idea for the speedup is to instead of searching for the next access
slot, have a lookup table that immediately tells where the next slot is for
each possible cycle (within a display line). The total size of this LUT is about
150kB, though each command only accesses a few chunks of ~3kB continuous
memory, so it should be OK for memory cache behaviour (and benchmarks confirm
it's indeed significantly faster).