Copied write-to-unused-registers behavior from Burczynski core
Writes to YM2413 registers 0x[123][9ABCDEF] actually end up 9(!) registers
lower. The Burczynski core implemented this from the start, even with a comment
that it's tested on real HW. Though because this is such weird behavior (from a
hardware point of view) I did re-test this myself (on a turbor machine).

So now it's also implemented for the Okazaki core. Though I doubt there exists
MSX software that actually makes use of this stuff.