Fixed (in a way) bug #498 Mouse emulation lacks arithmetic saturation
Bug #498 complains that mouse X/Y-offsets aren't clipped: if you move the mouse
very fast (and/or if the MSX program reads the mouse very slowly) then a large
positive offset wraps and becomes a (large) negative offset or vice versa. Note
that in practice this will happen _extremely_rarely_.

I verified this on real HW and a real MSX mouse behaves exactly in this way.
There is no clipping done in the mouse hardware. There actually was emulation
code to handle this clipping, but because of a bug it clipped on a too large
range (so by accident it already behaved more like the real HW than originally
intended). Fixed now by completely removing the clipping code.

Also fixed a spelling error (faze->phase) and added some more comments.