Basic touchpad support
It seems to work fine (only tested with PAD() in BASIC). Though the host-input
mechanism is very basic. It uses absolute mouse coordinates with (0,0) being
the top-left corner of the openmsx window. Instead we should try to map the
area so that the touchpad coordinates coincide with the MSX pixel positions.
But that's something for a later patch (I'm not actively working on it). Of
course having some touch host input device would be even better.