Added setting 'touchpad_transform_setting`
This allows to specify a 2x3 matrix that transforms the host-mouse-coordinates
to MSX-touchpad-coordinates. More in detail:
   | touch-X |    | a b c |   | host-X |
   | touch-X | =  | a b c | * | host-Y |
                              |    1   |
This allows to either:
 - allow to access the full 256x256 touchpad resolution
 - align touchpad to msx pixels
 - .. any (affine) transformation you can think of (rotate, scale, translate,
   shear)

The default matrix {{256 0 0} {0 256 0}} simply maps the full host window
to the full touchpad range. TODO we should provide some convenience Tcl scripts
that setup this matrix for various MSX applications (e.g. so that touchpad
input is aligned to MSX pixels).