Do not compile with clock_gettime on MinGW.
Newer versions of MinGW cross compilers (in Debian) do pass the
HAVE_CLOCK_GETTIME test, but do not implement it in the same library as
what we're using now. When everyone is using this version of MinGW we
should migrate to the library that implements it, but for now, let's
just revert to the SDL method as we did before.