Small cppcheck warning fix
Cppcheck warned that base and derived class both had a member with the same
name. In the derived classes IDECDROM and SCSIHD the member from the derived
class could be removed and instead add a (protected) getter on the base class.
In the base class VideoLayer, the member wasn't used at all and thus it can be
removed.