Fix auto-run cas from command line / remove debug command validation
Some background info:
  When setting a breakpoint (or watchpoint or ..) it's possible to specify a
  Tcl condition and/or a Tcl command that will be executed when the breakpoint
  hits. For example:
    proc my_custom_action {} { ... }
    set my_value ...
    debug set_bp 0x1234 {[vdpreg 1] == $::my_value} my_custom_action
  The debug command will do some basic validation on the expression and on the
  command. For example it will check some basic syntax errors, it will also
  check that the variable 'my_value' and the commands 'vdpreg' and
  'my_custom_action' exist. Though not all errors can be checked (and it's even
  possible the variables or procs disappear between the time-of-check and the
  time-of-use).

  In the world of static languages (e.g. C++) it's common to do checking as
  early as possible, preferably during compile-time. Though when using a
  dynamic language (like Tcl) the opposite is true, checking is only done right
  before the stuff is actually executed (and this sometimes gives more
  flexibility or enables some extra possibilities). So the early checks done by
  the debug command are not very Tcl-ish. All native Tcl commands that take
  callbacks don't perform early checks. Also our 'after' or 'bind' commands
  take callbacks and they don't have such checks.

The command validation has already been the source of a few bugs. E.g. we allow
to lazily load Tcl scripts. It is possible to use such lazy procs in
debug-callbacks. So we had to teach the verifier to not complain about lazy
procs. (Lazy proc definitions are one example of flexibilty enabled by a
dynamic language). Now we see another (related) problem:

When passing a .cas file on the openMSX command line, we (when enabled) will
execute cas-auto-load instructions. The implementation of this auto-loading is
partly done in Tcl and it uses a 'debug' command. Since the lazy-proc stuff,
the command verifier itself is also (partly) implemented in Tcl (inside
init.tcl). But when parsing the command line, the init.tcl script is not yet
executed, so the verification fails.

We could move the execution of init.tcl earlier (but it was actually
intentionally put rather late in the startup sequence to avoid doing useless
work in case of -testconfig or -bash). Though, at this moment, to me, it seems
better to just drop the verification feature:
- It's not what is typically done in Tcl (and we already don't do it for
  'after' or 'bind').
- It limits flexibility.
- It's not possible to catch all errors in the verifier.
- It's rather complex (not easy to get bug free)