Don't write settings.xml with an empty top level tag name
When you remove _both_ the settings.xml file from the openMSX user- and
system-directories (you have an incomplete openMSX installation at that
point!!), and you then start/stop openMSX, it will create a new user
settings.xml file. But the top level tag will have an empty name, and this is
an invalid xml file (even though openMSX itself doesn't actually use the name
of that tag).

So while this situation should never occur in a valid openMSX installation,
it's also very easy to fix. (And when developing openMSX it's sometimes
useful to fiddle with the system directory).