Fixed crash on 'set cputrace on'
This regression was likely introduced in rev 8bfce0, but I didn't verify. The
problem was that MSXCPU subscribed to cputrace setting changes, and passed
those to CPUCore. But CPUCore wasn't prepared to deal with them. A similar
problem occured for the {z80,r800}_freq{,_locked} settings on a turbor machine.

Although MSXCPU was subscribed to cputrace changes, it didn't actually do
anything on that event. So one possible fix was to simply not subscribe.
Instead this patch keeps the subscription and caches the value of the cputrace
setting in the callback (likely this was the intention all along).

Before this patch the CPUCore code directly queried the cputrace value from the
setting object, now it uses a cached value. Before revision dffe5f the value of
a BooleanSetting was stored directly (as a bool) in the object, so querying was
literally only 1 asm instruction. Now (all) setting values are stored in
TclObjects. Getting a boolean from a TclObjects is cheap (e.g. it will not
repeatedly perform the string->bool conversion), but it certainly is a lot more
than 1 asm instruction. The CPUCore code is careful to keep cputrace checks out
of the fast code path, but it still typically queries this setting ~100 times
per second and in certain scenarios possibly a factor 10x more (though likely
not 100000 times per second). So for performance using a cached version seemed
safer.