Make __restrict declarations match implementations
The GCC documentation states:

"As with all outermost parameter qualifiers, __restrict__ is ignored
in function definition matching. This means you only need to specify
__restrict__ in a function definition, rather than in a function
prototype as well."
  http://gcc.gnu.org/onlinedocs/gcc/Restricted-Pointers.html

However, it seems Clang (tested with version 3.3 on Linux) does require
the __restrict qualifier to be present on both: it issues "no matching
member function" errors if the prototype and the definition are out of
sync.

Also, Clang refuses to call non-restricted methods from within
restricted methods.