Fix(?) compilation error on vc++
This is a follow-up on revision 6b193d21. Clang seems to require __restrict
qualifiers on both declaration and definition. Gcc only needs it on the
definition, but accepts (and ignores) it also on the declaration. Visual studio
doesn't understand __restrict on the declaration. At least for restricted this-
pointers on member-function-pointers.

Strictly speaking C++ doesn't have a restrict keyword (that's why it's spelled
with leading double underscore), only the C99 standard has restrict. Restricted
this-pointers don't exist in any (C++) standard. So it's not surprising that
this extension doesn't work exactly the same in all compilers.

In this specific case it's not very clear whether having a restricted this-
pointer is a real performance win. I did verify that the generated code looks
different, but I couldn't easily tell (from looking at the generated asm code)
which variant was better. So in this case, for better portability, it's
probably best to just remove the __restrict annotations of the this-pointer.