Fixed assert on replaying a particular .omr file
'motion' managed to create a replay file which has a reset command that falls
in the middle of a Z80 instruction. So far it's unclear how exactly he managed
to create such a file. One possibility to create such files is to record/replay
using two openMSX versions which have different emulation timing and/or
behavior. But again, it's not clear whether this is indeed the case for
motion's replay.

In any case we need to be robust against loading replays that were created using
older openMSX versions. So I think this commit is useful anyway.

See comments in code for more details.