Restored Setting::setDontSaveValue() mechanism
This mechanism was removed in revision dffe5f98a: "Big setting class
refactoring". Re-added it and documented in which scenario it was useful.
So see code for more details.

Possibly there are better (=cleaner) ways to implement the unusual renderer
setting behavior at startup. This patch only restores the old behavior.