Fixed overflow in calculation in fast resampler
When running 'sme3promo' with resampler=fast you'll hear a lot of 'cracks'.
It works fine when using the blip or hq resampler.

The SCC soundchip occasionally produces high-amplitude spikes in its output
(outside the range [-32768,32767]). And because these values are later
multiplied with 16-bit fractional values, this can cause 32-bit integer
overflow in that calculation. Fixed by using only 14-bit fractional values,
gives us 2 bits extra room for amplitude spikes. 14 bit should still be
accurate enough to keep track of the fractional position between two samples,
even when this position is accumulated over ~1000 samples (the sample buffer
size). It's of course possible to perform these calculations using wider
integer types (or even floating point), but all those solutions are less suited
for a fast/low-quality resampler.

Unfortunately even with this fix there are still about as many cracks as before
in 'sme3promo'. I also investigated the rest of the audio-path but I didn't
find any more overflows in the calculations. My current best *guess* is that
these cracks occur in the higher frequency range that is filtered away by the
higher quality resamplers. And possibly this effect is amplified by the
aliasing artifacts introduced by the improper resampling (it is a
fast/low-quality resampler after all).

Anyway, this is still worth investigating further, but it's independent from
the fix in this patch.