Remove reversePool member from FilePool cache
The main motivation for this change is to simplify the modification described
in the next patch.

FilePool keeps a mapping between sha1sums and the corresponding filename.
Usually filepool is used to lookup a file based on a sha1. But sometimes we
also do the reverse: lookup the sha1 that belongs to a certain file. Before
this patch we had 2 structures that were kept in sync:
  pool: the main structure, sorted on sha1
  reversePool: sorted on filename, refers to entries in pool
So both type of searches could be done O(log(n)).

This patch removes the 'reversePool' member. So now lookups based on filename
are O(n). I don't believe this is bad for the following reasons:
 - Filename lookups aren't that frequent. And when we do lookup by filename
   it's often because we're scanning new files (and that part is much slower
   anyway).
 - Often right after a filename lookup, we will insert/remove an entry and
   because the underlying data structure is a std::vector that operation is
   also O(n). So basically the whole operation was already O(n).
 - There are several thousands of entries in the filepool, but not hundreds
   of thousands. So linear search is not too bad.
 - The reversePool takes a significant amount of memory, and keeping it
   up-to-date in parallel with the normal pool has some overhead. So removing
   reversePool might actually speedup the more common cases.
 - If later these filename searches turn up high in a profile-run (I doubt
   they will) we can always re-add a similar mechanism.