Fix iterator-invalidation bug in FilePool
The function getFromPool() will lookup a file for a given sha1sum. It's
possible there are multiple entries that match this sha1 (e.g. when the same
file is present in multiple locations in the filesystem). In such case the
getFromPool() function will loop over the matching entries, examine them one by
one, and return the first viable candidate.

Possible reasons why an entry with matching sha1 is not viable are:
 - file no longer exists
 - file has been modified (has a different sha1 now)
In both cases the getFromPool() routine would update the file pool with more
up-to-date information. That is: either remove the invalid entry or update the
sha1 and re-sort (remember: pool must be sorted on sha1). But this remove or
sort operation will invalidate iterators and because getFromPool() is still
looping over entries (using iterators) there is a bug, possibly resulting in a
crash.

The solution is to use indices instead of iterators to loop over the matching
entries (and then carefully increment or not increment the indices on remove
and sort operations).

I also optimized the re-sort operation: before when an entry's sha1 had
changed, we would remove the old entry and re-insert a new one. Instead it's
more efficient to keep the entry update it's content and then 'rotate' it to
the correct new position.

This regression was introduced in revision 938dbbbd9. There we changed the
underlying data structure from std::map to std::vector. Insert/remove in a map
does not invalidate iterators, and part of this code still made use of that
property.