Added 'too_fast_vram_access' setting
Possible values are 'real' and 'ignore'. When set to 'real' too fast VRAM
access results in dropped requests, like on a real machine. When set to
'ignore' timing is ignored, so all requests are executed. This might be useful
in combination with an over-clocked Z80.

Test-scenario:
  10 PRINT "Hello, world!"
  20 GOTO 10
This will result in a scrolling text, and internally the scroll routine does
a lot of VRAM read/writes. Now over-clock the Z80:
  set z80_freq_locked false
  set Z80_freq [expr 3579545 * 2]
A Z80 on 7MHz with the standard BIOS scroll routine will result in VRAM
corruption. Now enable the new setting
  set too_fast_vram_access ignore
And the scroll works fine again.

When emulating turbo-Z80 machines it's also required to enable this setting.
Those machines do have an extra hardware circuit that automatically slows down
the Z80 when it would otherwise access the VDP too fast, but those extra
circuits aren't emulated yet.