Ran include-what-you-use and removed some #includes
Include-what-you-use (IWYU) is a tool build on top of clang. It automatically
analyses which #includes should be added or can be removed from c++ files. IWYU
is very strict in #including everything what is used in a file without relying
on transitive #includes. Though in openmsx code we have the habbit of not
including something that obviously must already be included via some base
class. So in this patch I mostly only applied the include-removal part from
the IWYU output.

Netto this patch removes over 270 lines of code. But more important, it removes
a whole lot of #include dependencies, so incremental compilation should be
quicker now.

There might be some (minor) fallout on platforms I couldn't test-compile
(windows, mac).