Print error instead of crash on write error during video recording
On 32-bit systems the maximum file size is limited to 2GB (signed 32-bit
value). When recording a large .avi it's possible to reach that limit.
Before this patch this would result in a unhelpful warning message
  warning: Videorecording stopped, because you changed machine or
           changed a video setting during recording.
followed by a crash
  Uncaught exception: Error writing file
Now we should print the file writing error and not crash.

Thanks to BiFi for reporting this bug.

Ideally we should extend openmsx so that it can handle >2GB files also on
32-bit systems. Unfortunately that's a lot of work (use non-standard system
calls for file handling (on windows), but also (likely more work) audit many of
our internal data structures to make sure they can handle >4GB in 32-bit mode).
Also 32-bit systems are quickly disappearing, at least on the desktop, and
there you're most likely to encounter such large files. So implementing >2GB
file support on 32 bit systems is low priority for me (but I'd gladly accept
patches ;-).